requirements.txt
autopep8==1.6.0
click==8.1.3
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
grpcio==1.46.3
grpcio-tools==1.46.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
Pillow==9.1.1
protobuf==3.20.1
pycodestyle==2.8.0
PyYAML==6.0
six==1.16.0
toml==0.10.2
Werkzeug==2.1.2
zipp==3.8.0
